github_rag_engine/requirements.txt
streamlit
chromadb
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-llms-openai
python-dotenv
tiktoken
